\name{fire}
\alias{fire}
\docType{data}
\title{dataset}
\description{dataset to be used for some applications}
\usage{data("fire")}
\format{
  A data frame with 15 observations on the following 2 variables.
  \describe{
    \item{\code{DISTANCE}}{a numeric vector}
    \item{\code{DAMAGE}}{a numeric vector}
  }
}
\details{no further details}
\source{this dataset was taken from course materials}
\references{Mendenhall, William M., and Terry L. Sincich. 2016. Simple Linear Regression. 6th ed. Florida, USA: Taylor & Francis Group, LLC.}
\examples{
data(fire)
## maybe str(fire) ; plot(fire) ...
}
\keyword{datasets}
