% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Creating sample and corresponding table and bar graph}
\usage{
mybin(iter = 100, n = 10, p = 0.7)
}
\arguments{
\item{iter}{integer, number of iterations}

\item{n}{integer, sample size}

\item{p}{integer, percentage of success}
}
\value{
a table and bar graph corresponding to the sample
}
\description{
Creating sample and corresponding table and bar graph
}
\examples{
mybin(iter = 200, 10, 0.8)
}
