% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Title}
\usage{
myboot2(x, fun = "mean", alpha = 0.05, cx = 1.5, iter = 1000, ...)
}
\arguments{
\item{x}{size of sample}

\item{fun}{type of function}

\item{alpha}{percentage of the bootstrap interval}

\item{cx}{parameter that controls the size of text and symbols in graphs}

\item{iter}{number of iteration}

\item{...}{additional parameters}
}
\value{
creates a graph given the parameters
}
\description{
Title
}
\examples{
function(x,fun="mean",alpha=0.05,cx=1.5,...)
}
