% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{MyCLT function that creates a sample and a graph to represent it}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{size}

\item{iter}{number of iterations}

\item{a}{lower limit}

\item{b}{uppper limit}
}
\value{
histogram and a curve of the normal
}
\description{
MyCLT function that creates a sample and a graph to represent it
}
\examples{
myclt(n=10,iter=10000)

}
