% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum Likelihood function}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{function used for the matrix}

\item{x}{number of rows in the matrix created}

\item{param}{parameter values}

\item{...}{additional inputs/arguments}
}
\value{
a list
}
\description{
Maximum Likelihood function
}
\examples{
lfun = function(x, theta)
mymaxlik(lfun, x = 100, param=100)
}
