% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Create my N curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{value}
}
\value{
the curve
}
\description{
Create my N curve
}
\examples{
function(mu = 10, sigma = 10, a = 10)
}
