% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot.R
\name{myplot}
\alias{myplot}
\title{Quadratic function for a lab 4 task}
\usage{
myplot(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector
}
\description{
Quadratic function for a lab 4 task
}
\examples{
myplot(x = 10)
}
