% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Title}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats}

\item{gamma}{probability of a plane being overbooked}

\item{p}{probability of a "show"}
}
\value{
prints a named list containing nd, nc, N, p and gamma;
        creates a plot of Objective function Vs n
}
\description{
Title
}
\examples{
ntickets(N=400,gamma = 0.02, p = 0.95)

}
